Renamed file GODrawStop to GODrawstop
