+++ src/grandorgue/control/GOPistonControl.cpp
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2024 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2025 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -12,7 +12,7 @@
 
 #include "config/GOConfigReader.h"
 #include "model/GOCoupler.h"
-#include "model/GODrawStop.h"
+#include "model/GODrawstop.h"
 #include "model/GOManual.h"
 #include "model/GOOrganModel.h"
 #include "model/GOStop.h"
