+++ src/grandorgue/model/GOSwitch.h
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2024 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2025 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -8,7 +8,7 @@
 #ifndef GOSWITCH_H
 #define GOSWITCH_H
 
-#include "GODrawStop.h"
+#include "GODrawstop.h"
 
 class GOSwitch : public GODrawstop {
   // the number of manual for mainual switches
